feat: apply 3-layer generic hierarchy to TimePicker and TimeRangePicker
- TimePickerBase: abstract non-generic base (style-key, IClearControl, interaction logic)
- TimePickerBase<T>: generic intermediate (SelectedTime property, TTimeOnly converters)
- TimePicker: concrete TimePickerBase<TimeSpan>, StyleKeyOverride=typeof(TimePickerBase)

- TimeRangePickerBase: abstract non-generic base (style-key, IClearControl, interaction logic)
- TimeRangePickerBase<T>: generic intermediate (SelectedStart/EndTime properties, converters)
- TimeRangePicker: concrete TimeRangePickerBase<TimeSpan>, StyleKeyOverride=typeof(TimeRangePickerBase)

Update TimePicker.axaml and TimeRangePicker.axaml to target non-generic base types,
use IClearControl for Clear binding, and reference PART_ constants from base classes.

Fix ToTimeOnly(TimeSpan?) to use TimeOnly.FromTimeSpan() instead of missing extension.
